requirements.txt
requests
pytest
psycopg2-binary
python-dotenv
flask